requirements.txt
pypdf
svglib
qrbill
weasyprint
django-allauth
django-htmx
django-filter
django